change interface
the processing sketch now sends the data in blocks 6 bytes at a time followed by "\n".
Sending "\n" flushes the socket of the tcl server.
Changes on the vhdl code:
The signal udr of the virtual_jtag module is high for one tck after all data preceeding the "\n"
have been received.
This udr signal is used to write the data from the jtag data shift register into the memory.