stop simulation after one run
when no signals change any more than the simulation ends.
The signals stop changing when all processes are in an endless "wait;" status.
The clock is now generated inside a process. This process stops toggeling the clk
when stop_clk is set to '1'.
This is done at the end of the stimuli process.